+++ tests/index.test.js
@@ -1,6 +1,43 @@
 const createDisplay = require('..');
 
 describe('test', () => {
+  it('rounding function', () => {
+    const maxDoubleLength = 16;
+    const rounding = (intStr, decimalStr, decimalLength, type) => {
+      if (decimalStr.length <= decimalLength) {
+        return [intStr || '', decimalStr || ''];
+      }
+      if (intStr.length + decimalLength > maxDoubleLength) {
+        return [intStr || '', (decimalStr || '').slice(0, decimalLength)];
+      }
+      decimalLength = Math.max(decimalLength, 0);
+      const handler = type === 'ceil' ? Math.ceil : (type === 'floor' ? Math.floor : Math.round);
+      const rstStrs =
+        String(handler(`${intStr}.${decimalStr}e${decimalLength}`) / Math.pow(10, decimalLength)).split('.');
+      return [rstStrs[0] || '', rstStrs[1] || ''];
+    }
+
+    expect(rounding('123', '456', 2, 'round')).toStrictEqual(['123', '46']);
+    expect(rounding('123', '756', 0, 'round')).toStrictEqual(['124', '']);
+    expect(rounding('123', '456', 2)).toStrictEqual(['123', '46']);
+    expect(rounding('123', '456', 5)).toStrictEqual(['123', '456']);
+    expect(rounding('123', '9999', 3)).toStrictEqual(['124', '']);
+    expect(rounding('123', '001', 2)).toStrictEqual(['123', '']);
+    expect(rounding('123', '005', 2)).toStrictEqual(['123', '01']);
+
+    expect(rounding('123', '9999', 3, 'floor')).toStrictEqual(['123', '999']);
+    expect(rounding('123', '9999', 0, 'floor')).toStrictEqual(['123', '']);
+    expect(rounding('123', '005', 2, 'floor')).toStrictEqual(['123', '']);
+
+    expect(rounding('123', '4501', 2, 'ceil')).toStrictEqual(['123', '46']);
+    expect(rounding('123', '4511', 2, 'ceil')).toStrictEqual(['123', '46']);
+    expect(rounding('123', '45', 2, 'ceil')).toStrictEqual(['123', '45']);
+    expect(rounding('123', '01', 0, 'ceil')).toStrictEqual(['124', '']);
+
+    expect(rounding('123', '756', -1, 'round')).toStrictEqual(['124', '']);
+    expect(rounding('123', '9999', -10, 'floor')).toStrictEqual(['123', '']);
+  });
+
   it('default display', () => {
     const display = createDisplay();
 
@@ -20,18 +57,18 @@ describe('test', () => {
     expect(display('12..34')).toBe('12..34');
 
     expect(display(-123456789.123456789)).toBe(display('-123456789.123456789'));
-    expect(display(-123456789.123456789)).toBe('-123.456M')
+    expect(display(-123456789.123456789)).toBe('-123.457M')
     expect(display(123456)).toBe('123,456');
-    expect(display(123.456)).toBe('123.45');
+    expect(display(123.456)).toBe('123.456');
     expect(display('123.4')).toBe('123.4');
-    expect(display('12345678.9')).toBe('12345678');
+    expect(display('12345678.9')).toBe('12345679');
 
     expect(display('123.0')).toBe('123');
     expect(display(123.0)).toBe('123');
     expect(display('123.000')).toBe('123');
     expect(display(123.000)).toBe('123');
-    expect(display(23.008)).toBe('23');
-    expect(display(1234567.08)).toBe('1234567');
+    expect(display(23.008)).toBe('23.008');
+    expect(display(1234567.08)).toBe('1234567.1');
     expect(display(100000000000)).toBe('100G');
 
     expect(display(-1.2345e+5)).toBe('-123,450');
@@ -46,10 +83,24 @@ describe('test', () => {
     expect(display(-.34)).toBe('-0.34');
   });
 
-  it('placeholder allowText comma', () => {
+  it('rounding', () => {
+    const displayRound = createDisplay({roundingType: 'round', precision: 2});
+    const displayFloor = createDisplay({roundingType: 'floor', precision: 2});
+    const displayCeil = createDisplay({roundingType: 'ceil', precision: 2});
+
+    expect(displayRound(23.008)).toBe('23.01');
+    expect(displayFloor(23.008)).toBe('23');
+    expect(displayCeil(23.001)).toBe('23.01');
+    
+    expect(displayRound(-123456789.123456789)).toBe('-123.457M');
+    expect(displayFloor(-123456789.123456789)).toBe('-123.456M');
+    expect(displayCeil(-123456189.123456789)).toBe('-123.457M');
+  });
+
+  it('placeholder allowText separator', () => {
     const display = createDisplay({
       allowText: false,
-      comma: false,
+      separator: false,
       placeholder: '--'
     });
 
@@ -62,7 +113,7 @@ describe('test', () => {
 
     const displayText = createDisplay({
       allowText: true,
-      comma: false,
+      separator: false,
       placeholder: '--'
     });
 
@@ -76,7 +127,7 @@ describe('test', () => {
   it('length decimal', () => {
     const display = createDisplay({
       length: 6,
-      decimal: 10,
+      precision: 10,
     });
 
     expect(display(123456)).toBe('123456');
@@ -93,11 +144,11 @@ describe('test', () => {
   it('int', () => {
     const display = createDisplay({
       length: 6,
-      decimal: 0,
+      precision: 0,
     });
 
-    expect(display(1234.8)).toBe('1,234');
-    expect(display(123456.8)).toBe('123456');
+    expect(display(1234.8)).toBe('1,235');
+    expect(display(123456.8)).toBe('123457');
   });
 
   it('length too short', () => {
@@ -108,6 +159,6 @@ describe('test', () => {
 
     expect(display(null)).toBe('abc');
 
-    expect(() => display(-123456)).toThrowError('length: 3 is too small');
+    expect(display(-123456)).toBe('-123456');
   });
 });